Add --qtimpl commandline option
The efforts to select which Qt Implementation to use at runtime have
been mostly successful, but even thought PyQt4 and PyQt5 can coexist on
the same system, they can't coxexist in the same process.

For that reason a --qtimpl option has been added to py.test via
conftest.py, this switch lets you pick whether to test PySide, PyQt4 or
PyQt5.

Accompanying updates to tox configuration allow this to be switche out
as well.

The README will need to be updated to acknowledge this change.